Fixed Silicon-Wands issue #4.
The problem was that ghost operations such as (un)folding use consume to temporarily change the heap, which overwrote context.reserveEvalHeap with the current context.reserveHeap, but did not restore the previous reserveEvalHeap afterwards. This resulted in permissions no longer being available, because they had already been removed from the reserveHeap, but were still present in the overwritten reserveEvalHeap.