Fixed issues 27,29,39,44
Added traits to the parser hierarchy and reshuffled such that
global and local declarations are differentiated,
all named declarations inherit the same trait and
so each declaration now goes to the correct scope (some where missed before).
Added axioms to the global scope so that name clashes are reported in time.

Added the error id to  unexpected error messages.

Added purity checks for old expressions and repaired them for
function applications.

Disallowed ++ for sequences.

Added test cases for issues (44 is related to 27)