Removed the DefaultFractionalPermissions, and P <: FractionalPermissions type parameters.
The first reason was that the type safety implied by working with FractionalPermissions/DefaultFractionalPermissions was not actually given, because TermPerms(t) could be used to wrap any Term t, thereby turning it into a DefaultFractionalPermissions. It is actually necessary to use arbitrary terms where permissions are expected because of permission-typed variables and functions. Hence, requiring FractionalPermissions did not actually statically guarantee that a permission-typed term is passed; an additional runtime check inside TermPerms(t) was still necessary.

The second reason for removing type parameters of the kind P <: FractionalPermissions was that the last three years showed that only a single permission representation is used at any time, i.e., that Silicon is not used with two different representations at the same time. The same holds for heaps, states and contexts, and I plan to remove type parameters for those as well (and use a "universe" with abstract type members instead).