Changed assignments (to local variables and fields) to use SSA-form in cases where the right-hand side contains expressions that are not allowed in triggers. This was necessary because the previous approach resulted in invalid triggers. Given the snippet   var pw: Seq[Int] := [0..n+1)   inhale forall i: Int :: {pw[i]} ... the trigger expression pw[i] was previously translated (evaluated) to   [0..n+1)[i] which is not a valid trigger (because of the arithmetic addition).
Using SSA-form assignments turned out to slow down the verification notably (from 3min:25sec to 4min:10sec), so it is only used when the right-hand side contains expressions that are not allowed in triggers, which bringes to verification time down to 3min:30s again.

Conveniently, using SSA-form assignments is also a way of (somewhat indirectly) fixing Silicon issue 110.