Follow-up to commit 353b511d058d: abstract over field value functions (snapshot maps) in triggers as well. This is necessary in situations where the definition of a fvf is nested under a quantifier whose trigger mentions that particular fvf. For example, given the Viper quantifier
  forall x: Ref :: {x.f} ... ==> x.f == 0

where permission to (all) x.f come from quantified permission, an auxiliary quantifier of the shape

  forall x :: {fvf1(x)}
       ...
    && forall r: Ref :: {...} fvf1(r) == fvf0(r)  // fvf0 is the FVF of the chunk providing permission to x.f

will be generated (in addition to the actual quantifier forall x: Ref :: {fvf1(x)} ... ==> fvf1(x) == 0

Previously, it could happen that the solver needed to know the definition of fvf1, e.g. to know that fvf1(y) == fvf0(y), but it only had fvf2(y) as a trigger and the knowledge that fvf2(y) == fvf0(y). However, none of these terms would allow the solver to instantiate the auxiliary quantifier from above.

A more restrictive (and thus potentially performance-improving) alternative would be "expand" triggers that mention an fvfN to all other fvf's relative to which fvfN is defined. E.g. the above auxiliary quantifier would become

  forall x :: {fvf1(x)} {fvf0(x)} ...