Added specialised axioms for indexing into the concatenation of some sequence and the empty sequence. This seems to prevent a rather weird Z3 matching loop which results in instantiation chains of the following shape (let AX denote the previous axiom for index over append):
     (Nil ++ s)[j]
  == s[j + (-1')*|Nil|]  // AX
  == (Nil ++ s)[j + (-1')*|Nil|]
  == (Nil ++ s)[j + (-1')*|Nil| + (-1'')*|Nil|]  // AX
  == (Nil ++ s)[j + (-1'')*|Nil|]
  == (Nil ++ s)[j + (-1'')*|Nil| + (-1''')*|Nil|]  // AX
  ...

Each "-1" term (and thus the nesting "+" terms) had a different term identifier (in the Z3 log output), which seems to fuel the matching loop. It is not clear, however, why the newly-added and/or rewritten axioms help, though. However, the previously observed matching loop (exhibited when running Silicon on our online examples linked-list-predicates and linked-list-wands) disappeared with this change.